Playing around with stuff and refactoring
